refactor(theme): move utility file
