+++ packages/vuetify/src/services/theme/__tests__/theme-utils.spec.ts
@@ -1,7 +1,7 @@
 import {
   genStyles,
   parse
-} from '../theme'
+} from '../utils'
 
 describe('theme-utilities.ts', () => {
   let parsedTheme
